pyproject.toml
[tool.poetry]
name = "agent-memory-system"
version = "0.1.0"
description = "A memory management system for intelligent agents with dual-track memory mechanism"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "agent_memory_system"}]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.0.0"
transformers = "^4.30.0"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
networkx = "^3.1"
neo4j = "^5.9.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
pydantic = "^2.0.0"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^2.17.0"
mypy = "^1.4.0"
pre-commit = "^3.3.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.pylint]
max-line-length = 88
disable = ["C0111", "C0103"]

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true 